Lowered cost of base upgrades, nerfed Turrets, Silver and Vamp, Adjusted PercentageToWin numbers

+++ devtools/data-editor/index.js
@@ -156,7 +156,7 @@ function buildUIFromData(data, schema, layer = 0, onChange = (val) => {}) {
     if (type === 'number') {
         const input = $(`<input type="number" value="${data}" />`);
         input.change(() => {
-            onChange(parseInt(input.val()));
+            onChange(parseFloat(input.val()));
             updateDiff();
         });
         return input;
